vanna-service/requirements.txt
# Vanna Text2SQL 服务依赖
openai>=1.0.0
chromadb>=0.4.0
pandas>=2.0.0
pymysql>=1.1.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
sqlparse>=0.4.0
plotly>=5.0.0
